+++ .rubocop.yml
@@ -24,7 +24,7 @@ Layout/BlockAlignment:
 
 Metrics/ParameterLists:
   Exclude:
-    - 'lib/stoplight/config.rb'
+    - 'lib/stoplight/light/config.rb'
 
 Style/CaseEquality:
   Enabled: false
